Move field_shorthand to a separate module
